misc/requirements.txt
azure-mgmt-resource>=18.0.0
azure-mgmt-compute
azure-mgmt-network
azure-identity>=1.5.0